# [Thinkit] Improve performance of Bazel Test Environment by caching open file descriptors, Add a benchmark for append time given a fixed size file, Add SetTestCaseIDs to TestEnvironment to support setting multiple IDs & Add some clarifying documentation to generic interface files.